+++ src/lib/AsyncTypeaheadInput.tsx
@@ -173,7 +173,7 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadInputPr
         onChange={(_e, value) => {
           // value is typed as Autocomplete<Value> (aka TypeaheadOption) or an array of Autocomplete<Value> (aka TypeaheadOption[])
           // however, the component is not intended to be used with mixed types
-          const optionsArray = value ? (Array.isArray(value) ? value : [value]) as TypeaheadOptions : undefined;
+          const optionsArray = value ? ((Array.isArray(value) ? value : [value]) as TypeaheadOptions) : undefined;
           setValue(optionsArray ?? []);
           const values = convertAutoCompleteOptionsToStringArray(optionsArray);
           const finalValue = multiple ? values : values[0];